test: add app tests

+++ src/fileServices/s3Service.ts
@@ -4,6 +4,7 @@ import contentDisposition from 'content-disposition';
 import { GraaspS3FileItemOptions } from '../types';
 import FileService from './interface/fileService';
 import { S3FileNotFound } from '../utils/errors';
+import { StatusCodes } from 'http-status-codes';
 
 export class S3Service implements FileService {
   private readonly options: GraaspS3FileItemOptions;
@@ -27,40 +28,6 @@ export class S3Service implements FileService {
         useAccelerateEndpoint,
         credentials: { accessKeyId, secretAccessKey },
       });
-  }
-
-  async getMetadata(key: string) {
-    const { s3Bucket: Bucket } = this.options;
-    const metadata = await this.s3Instance
-      .headObject({ Bucket, Key: key })
-      .promise();
-    return metadata;
-  }
-
-  // get file buffer, used for generating thumbnails
-  async getFileBuffer({ filepath }): Promise<Buffer> {
-    const { s3Bucket: bucket } = this.options;
-    const params = {
-      Bucket: bucket,
-      Key: filepath,
-    };
-    return (await this.s3Instance.getObject(params).promise()).Body as Buffer;
-  }
-
-  async downloadFile({ reply, filepath, itemId }) {
-    const { s3Bucket: bucket, s3Region: region } = this.options;
-    try {
-      // check whether file exists
-      await this.getMetadata(filepath);
-
-      // Redirect to url, TODO: Change for something better
-      reply.redirect(
-        `https://${bucket}.s3.${region}.amazonaws.com/${filepath}`,
-      );
-    } catch (e) {
-      // TODO: check error and return the corresponding one
-      throw new S3FileNotFound({ filepath, itemId });
-    }
   }
 
   async copyFile({
@@ -103,6 +70,7 @@ export class S3Service implements FileService {
       .promise();
   }
 
+  // delete all content in a folder
   async deleteFolder({ folderPath }): Promise<void> {
     const { s3Bucket: bucket } = this.options;
 
@@ -114,17 +82,62 @@ export class S3Service implements FileService {
 
     const filepaths = Contents.map(({ Key }) => Key);
 
-    await this.s3Instance
-      .deleteObjects({
-        Bucket: bucket,
-        Delete: {
-          Objects: filepaths.map((filepath) => ({
-            Key: filepath,
-          })),
-          Quiet: false,
-        },
-      })
+    if (filepaths.length) {
+      await this.s3Instance
+        .deleteObjects({
+          Bucket: bucket,
+          Delete: {
+            Objects: filepaths.map((filepath) => ({
+              Key: filepath,
+            })),
+            Quiet: false,
+          },
+        })
+        .promise();
+    }
+  }
+
+  async downloadFile({ reply, filepath, itemId }) {
+    const { s3Bucket: bucket, s3Region: region } = this.options;
+    try {
+      // check whether file exists
+      await this.getMetadata(filepath);
+
+      // Redirect to url, TODO: Change for something better
+      reply.redirect(
+        `https://${bucket}.s3.${region}.amazonaws.com/${filepath}`,
+      );
+    } catch (e) {
+      if (e.statusCode === StatusCodes.NOT_FOUND) {
+        throw new S3FileNotFound({ filepath, itemId });
+      }
+      throw e;
+    }
+  }
+
+  // get file buffer, used for generating thumbnails
+  async getFileBuffer({ filepath }): Promise<Buffer> {
+    const { s3Bucket: bucket } = this.options;
+    const params = {
+      Bucket: bucket,
+      Key: filepath,
+    };
+    try {
+      return (await this.s3Instance.getObject(params).promise()).Body as Buffer;
+    } catch (e) {
+      if (e.statusCode === StatusCodes.NOT_FOUND) {
+        throw new S3FileNotFound({ filepath });
+      }
+      throw e;
+    }
+  }
+
+  async getMetadata(key: string) {
+    const { s3Bucket: Bucket } = this.options;
+    const metadata = await this.s3Instance
+      .headObject({ Bucket, Key: key })
       .promise();
+    return metadata;
   }
 
   async uploadFile({
